--- a/src/Rx.KitchenSink.ts
+++ b/src/Rx.KitchenSink.ts
@@ -149,6 +149,9 @@ observableProto.groupBy = groupBy;
 import isEmpty from './operators/isEmpty';
 observableProto.isEmpty = isEmpty;
 
+import every from './operators/every';
+observableProto.every = every;
+
 import last from './operators/last';
 observableProto.last = last;