adding `every` operator.

Micro perf test shows noticeable slowness on new operator.

>old every(predicate) with immediate scheduler x 104,692 ops/sec ±1.20% (95 runs sampled)
>new every(predicate) with immediate scheduler x 87,236 ops/sec ±1.24% (90 runs sampled)
>        -16.67% slower than Rx2

>old every(predicate, thisArg) with immediate scheduler x 94,638 ops/sec ±1.37% (96 runs sampled)
>new every(predicate, thisArg) with immediate scheduler x 84,714 ops/sec ±0.53% (96 runs sampled)
>        -10.49% slower than Rx2

investigating root cause but not yet able to find it, created PR to share implementation to address mistake / errors that I'm missing.