@kwonoj 2015-10-01
* updated couple of testcases to ensure source element before subscription doesn't affect behavior of `every`