@blesh 2015-10-01
@kwonoj ... I've found no such slowness.

> old every(predicate) with immediate scheduler x 100,478 ops/sec ±0.91% (90 runs sampled)
> new every(predicate) with immediate scheduler x 278,192 ops/sec ±0.75% (93 runs sampled)
	176.87% faster than Rx2

> old every(predicate, thisArg) with immediate scheduler x 91,583 ops/sec ±0.57% (92 runs sampled)
> new every(predicate, thisArg) with immediate scheduler x 247,939 ops/sec ±0.61% (95 runs sampled)
	170.73% faster than Rx2