@kwonoj 2015-10-01
Interesting, seems my system configuration should have some oddness related with micro perf slowness, recall history of observing slowness only on my end :(