@blesh 2015-10-01
It's the `Rx.config.slow_for_kwonoj = true` setting I added a month ago.