@blesh 2015-10-02
Made a change that detects scalars for an increase in perf:

was

> old every(predicate) with immediate scheduler x 321,042 ops/sec ±0.75% (95 runs sampled)
> new every(predicate) with immediate scheduler x 1,921,189 ops/sec ±0.70% (96 runs sampled)
>	498.42% faster than Rx2

now

> old every(predicate) with immediate scheduler x 327,967 ops/sec ±0.70% (93 runs sampled)
> new every(predicate) with immediate scheduler x 4,126,076 ops/sec ±0.67% (94 runs sampled)
> 	1158.08% faster than Rx2

and also I did the same for detecting observables of static arrays:

was

> old static array observable every(predicate) with immediate scheduler x 190,344 ops/sec ±0.48% (96 runs sampled)
> new static array observable every(predicate) with immediate scheduler x 1,895,897 ops/sec ±0.62% (95 runs sampled)
	896.04% faster than Rx2

now 

> old static array observable every(predicate) with immediate scheduler x 179,314 ops/sec ±0.66% (93 runs sampled)
> new static array observable every(predicate) with immediate scheduler x 4,134,918 ops/sec ±0.72% (92 runs sampled)
	2205.96% faster than Rx2

Merged with d5120d2

Thanks @kwonoj!