@blesh 2015-10-02
... in retrospect, I should have added those two new micro perf tests in separate previous commits... since i did them first... oh well. Next time.