@blesh 2015-10-02
FWIW, the change I made is something we could do in other places for sure. `isEmpty` or `first` for example. Anything that returns a ScalarObservable could be optimized most likely.